Gravity tool added 75%
Render function working at 90%. There is a bug for solving in further commits.

Notes:
Drawing functions improved to encapsulate just the related code of drawing.
Groups are now created in the function drawGraph.
The difference between the group center and the circle center is now set as a property: group.childDX and group.childDY.